feat(bdd): web UI mode — ferridriver bdd --ui
Playwright-UI-mode analogue for BDD suites: a self-contained localhost
web app served by the test runner. No build tooling, no frameworks, no
external requests; dark theme.

Server (crates/ferridriver-test/src/ui_server.rs):
- GET / serves the inline app; GET /ws streams JSON messages (testList,
  runStarted, testStarted, stepStarted/Finished, testFinished with
  outcome/attachments, runFinished totals, watchStatus) via a broadcast
  channel so N tabs stay in sync; lagged tabs resync from a snapshot.
- Client commands: runAll, runFailed, runGrep, runTest, runFile, stop.
- GET /artifact/<rel> serves files strictly under the run's output_dir
  (traversal + symlink escape rejected, unit-tested) with
  Access-Control-Allow-Origin: * so trace.playwright.dev can fetch
  trace zips. Attachment urlPaths are percent-encoded.
- Explicit ReporterEvent -> JSON mapping; internal types stay
  unserialized.

Runner (TestRunner::run_ui): same skeleton as run_watch — browser
launched once and reused, file watcher narrows feature edits, full plan
rebuild refreshes the sidebar. Trace mode forced to On when Off so
every test gets a trace attachment. Stop cancels the in-flight run by
dropping the execute future (TUI-cancel semantics); commands arriving
mid-run queue and run after. Binds 127.0.0.1, --ui-port or ephemeral,
URL printed on startup. --ui wins over --watch when both are passed.

Frontend: collapsible suite tree with live status dots and per-suite
counts, status filter chips, text filter (Enter = grep run), Run
All/Run Failed/Stop, per-test and per-file run buttons, progress bar,
keyboard navigation (arrows/Enter//), resizable sidebar, reconnecting
websocket; detail pane with Steps (nested, durations, inline errors),
Attachments (download + "Open in trace viewer" link), and Output
(stdout/stderr) tabs.

Runner step traces ported to trace format VERSION 8
(crates/ferridriver-test/src/tracing.rs): first line context-options
with version 8, merged action events with parentId nesting and stack
frames from step locations, trace.trace + trace.network zip entries —
the old emitter wrote versionless test.trace files the viewer
mis-modernized. Worker now passes the real browser name.

Tests: unit tests for the event mapping, artifact path guard, URL
encoding, command parsing, idle-sweep, and v8 trace invariants; E2E
test (tests/ui_mode.rs) spawns the built binary on a scratch feature
dir, connects over websocket, asserts testList, drives runAll, checks
testFinished/runFinished, fetches the trace over HTTP and validates
CORS + the v8 first line + traversal rejection. Verified live in a
real browser (Chromium via MCP) including failure rendering and
encoded artifact URLs.

Also fixed in this commit (pre-existing red under the current nightly):
16 clippy errors in the core crate (doc backticks, cast lints with
justified allows, new_page viewport resolution extracted to
effective_viewport, redundant closure, items-after-statements), 3 in
ferridriver-node (dead inner_ref removed, documented &Env allows), test
lint fixes in backends_support, and the sidecars gateway test still
passing --plugin after the plugin->extension rename (now --extension).

Known gaps (docs/PLAYWRIGHT-PARITY-BACKLOG.md updated): step traces
have no DOM snapshots, screencast frames, or network entries; step
timelines are reconstructed sequentially from durations;
screenshot-on-failure attachments are in-memory bytes with no download
link; no embedded trace viewer (viewer link needs network access);
ferridriver test (#[ferritest]) has no UI mode — bdd only.

+++ crates/ferridriver/src/tracing.rs
@@ -267,7 +267,7 @@ impl Tracing {
           // Approximate wire order: the log preserves arrival order but
           // per-request monotonic capture isn't wired yet, so entries
           // are spaced inside the chunk for stable viewer sorting.
-          obj.insert("_monotonicTime".to_string(), serde_json::json!(i as f64));
+          obj.insert("_monotonicTime".to_string(), serde_json::json!(i));
         }
         entries.push(value);
       }
@@ -311,12 +311,12 @@ pub(crate) async fn flush_recorder(recorder: &HarRecorder, requests: &[Request])
   let json = serde_json::to_string_pretty(&archive).map_err(|e| FerriError::backend(format!("serialize HAR: {e}")))?;
 
   if is_zip_path(&recorder.path) {
+    use std::io::Write;
     let file = std::fs::File::create(&recorder.path)
       .map_err(|e| FerriError::backend(format!("create HAR zip {}: {e}", recorder.path.display())))?;
     let mut writer = zip::ZipWriter::new(file);
     let opts = zip::write::SimpleFileOptions::default().compression_method(zip::CompressionMethod::Deflated);
     let zip_err = |e: zip::result::ZipError| FerriError::backend(format!("write HAR zip: {e}"));
-    use std::io::Write;
     writer.start_file("har.har", opts).map_err(zip_err)?;
     writer
       .write_all(json.as_bytes())
@@ -558,7 +558,7 @@ fn mime_extension(mime_type: &str) -> &'static str {
 }
 
 /// Case-insensitive header lookup — CDP lowercases header names but
-/// WebKit and BiDi deliver them as sent (`Content-Type`), and a HAR
+/// `WebKit` and `BiDi` deliver them as sent (`Content-Type`), and a HAR
 /// recorded with `mimeType: x-unknown` replays as a download instead of
 /// a rendered document.
 fn header_value(headers: &crate::network::Headers, name: &str) -> Option<String> {
